move `arch` directory back into `kernel`
