+++ kernel/include/kernel/arch/x86/devices/vga.h
