+++ kernel/src/arch/x86/32/gdt.c
@@ -1,5 +1,5 @@
-#include "arch/gdt.h"
-#include "arch/descriptor_table.h"
+#include "kernel/arch/x86/gdt.h"
+#include "kernel/arch/x86/descriptor_table.h"
 
 #define LIMIT_4KiB 0xfffffU
