+++ kernel/src/arch/x86/32/isr.c
@@ -1,6 +1,6 @@
-#include "arch/32/isr.h"
-#include "arch/32/idt.h"
-#include "arch/devices/vga.h"
+#include "kernel/arch/x86/32/isr.h"
+#include "kernel/arch/x86/32/idt.h"
+#include "kernel/arch/x86/devices/vga.h"
 
 #define ISR_HANDLER_ALIGNMENT      (16U)
 #define ISR_EXCEPTION_VECTOR_START (0x00U)
