+++ kernel/src/arch/x86/32/linker.ld
