+++ kernel/src/arch/x86/port.c
@@ -1,4 +1,4 @@
-#include "arch/port.h"
+#include "kernel/arch/x86/port.h"
 #include <stdint.h>
 
 uint8_t
