+++ kernel/src/kpanic.c
@@ -1,6 +1,6 @@
 #include "kernel/kpanic.h"
-#include <arch/cpu.h>
-#include <arch/devices/vga.h>
+#include "kernel/arch/x86/cpu.h"
+#include "kernel/arch/x86/devices/vga.h"
 
 void
 kernel_on_panic(const char *msg, const char *location, const char *function)
