+++ meson.build
@@ -1,4 +1,3 @@
 project('t5os', 'c', version: '0.1.0', meson_version: '>= 1.5.0')
 
-subdir('arch')
 subdir('kernel')
